+++ CHANGELOG.md
@@ -12,6 +12,8 @@ ENHANCEMENTS:
 * managed-redis: added 'notify_keyspace_events', 'slowlog_log_slower_than', 'slowlog_max_len' and 'databases' fields
 * add `maintenance_window` section in `yandex_mdb_clickhouse_cluster` resource and data source
 * add `num_partitions` and `default_replication_factor` attributes in `yandex_mdb_kafka_cluster` resource and data source
+* change of `dns_record`, `ipv6_dns_record` and `nat_dns_record` in `network_interface` entity of `yandex_compute_instance`
+without instance drop
 
 BUG FIXES:
 * mdb: throw error when trying to modify `owner` in `database` block in `yandex_mdb_postgresql_cluster`